+++ src/tests/likes.test.js
@@ -1,0 +1,1 @@
+import './modules/homepage.js';
